QR scanning function, albeit roughly.
